Progress on tests: Need to work on REST auth route.

+++ migrations/20180508194440_service_identifier.js
@@ -6,6 +6,10 @@ exports.up = function(knex, Promise) {
 
 exports.down = function(knex, Promise) {
   return knex.schema.table("services", t => {
-    t.dropColumn("service_identifier");
+    return knex.schema.hasColumn("services", "service_identifier").then(exists => {
+      if (exists) {
+        t.dropColumn("service_identifier");
+      }
+    });
   });
 };
